Rows and Zig-zag Moving
Add RowData struct and list for that objects, which are has all specific information about the Row.
Now code works with list of rows and information about current row, instead of just numbers.
Rewrite the moving up and down to zig-zag-like style based on specific max count of decoration in the row.